Create assignment directly when the create button is clickedd
