+++ client/src/router.config.tsx
@@ -3,7 +3,6 @@ import { AUTHORITIES } from './hooks/useHasAuthority'
 import AdminPage from './pages/AdminPage'
 import AssignmentListPage from './pages/assignment/AssignmentListPage'
 import AssignmentPage from './pages/assignment/AssignmentPage'
-import CreateAssignmentPage from './pages/assignment/CreateAssignmentPage'
 import {
   BasicHelpPage,
   DefinitionsHelpPage,
@@ -30,12 +29,6 @@ export const routerConfig: Route = {
       hideChildrenInMenu: true,
       component: AssignmentListPage,
       children: [
-        {
-          path: '/assignments/create',
-          name: 'Create Assignment',
-          authority: AUTHORITIES.ROLE_TEACHER,
-          component: CreateAssignmentPage
-        },
         {
           path: '/assignments/:id',
           component: AssignmentPage
